+++ CHANGELOG.md
@@ -3,3 +3,11 @@
 ## v0.1.0
 
 Initial release, the frogs have arrived
+
+## v0.1.2
+
+- Adds FrogDataContainerEasy for simple data management
+
+## v0.1.3
+
+- Adds AssetMUD for asset tracking in sidecar data